minimal-versions: run tests using `stable` Rust (#33)
Adds a configurable `toolchain` parameter to the reusable workflow for
minimal-versions which allows a different toolchain than `nightly` to be
used for the purposes of running tests.

`nightly` is only needed for the `-Z minimal-versions` resolution of
Cargo.lock. We can avoid nightly breakages of which we've hit quite a
few lately by reducing the scope of what we use `nightly` for to *just*
that version resolution.

Additionally, this moves the resolution of Cargo.lock to earlier in the
workflow so it applies to all steps. Previously benchmarks were being
tested before the `-Z minimal-versions` resolution took place.